# apps/earp-server/openapi.yaml
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    ManifestActivateRequest:
      additionalProperties: false
      properties:
        attestation:
          additionalProperties: true
          title: Attestation
          type: object
        expected_active_revision:
          anyOf:
          - minimum: 1.0
            type: integer
          - type: 'null'
          title: Expected Active Revision
        manifest_id:
          maxLength: 128
          minLength: 1
          title: Manifest Id
          type: string
        manifest_revision:
          minimum: 1.0
          title: Manifest Revision
          type: integer
        packs:
          items:
            $ref: '#/components/schemas/ManifestPackSelection'
          maxItems: 3
          minItems: 1
          title: Packs
          type: array
        profile_id:
          maxLength: 128
          minLength: 1
          title: Profile Id
          type: string
      required:
      - profile_id
      - manifest_id
      - manifest_revision
      - packs
      - attestation
      title: ManifestActivateRequest
      type: object
...
    ManifestPreviewRequest:
      additionalProperties: false
      properties:
        manifest_id:
          maxLength: 128
          minLength: 1
          title: Manifest Id
          type: string
        manifest_revision:
          minimum: 1.0
          title: Manifest Revision
          type: integer
        packs:
          items:
            $ref: '#/components/schemas/ManifestPackSelection'
          maxItems: 3
          minItems: 1
          title: Packs
          type: array
        profile_id:
          maxLength: 128
          minLength: 1
          title: Profile Id
          type: string
      required:
      - profile_id
      - manifest_id
      - manifest_revision
      - packs
      title: ManifestPreviewRequest
      type: object
...
info:
  title: EARP Server
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /v1/catalog/manifests/activate:
    post:
      operationId: activate_manifest_v1_catalog_manifests_activate_post
      parameters:
      - in: header
        name: Idempotency-Key
        required: true
        schema:
          maxLength: 128
          minLength: 1
          title: Idempotency-Key
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManifestActivateRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Activate Manifest V1 Catalog Manifests Activate Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Activate Manifest
      tags:
      - catalog
  /v1/catalog/manifests/preview:
    post:
      operationId: preview_manifest_v1_catalog_manifests_preview_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManifestPreviewRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Preview Manifest V1 Catalog Manifests Preview Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Preview Manifest
      tags:
      - catalog
  /v1/catalog/manifests/revoke:
    post:
      operationId: revoke_manifest_v1_catalog_manifests_revoke_post
      parameters:
      - in: header
        name: Idempotency-Key
        required: true
        schema:
          maxLength: 128
          minLength: 1
          title: Idempotency-Key
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManifestRevokeRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Revoke Manifest V1 Catalog Manifests Revoke Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Revoke Manifest
      tags:
      - catalog
  /v1/catalog/manifests/rollback:
    post:
      operationId: rollback_manifest_v1_catalog_manifests_rollback_post
      parameters:
      - in: header
        name: Idempotency-Key
        required: true
        schema:
          maxLength: 128
          minLength: 1
          title: Idempotency-Key
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManifestRollbackRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Rollback Manifest V1 Catalog Manifests Rollback Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Rollback Manifest
      tags:
      - catalog
...
  /v1/catalog/packs:
    get:
      operationId: packs_v1_catalog_packs_get
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties:
                  items:
                    additionalProperties: true
                    type: object
                  type: array
                title: Response Packs V1 Catalog Packs Get
                type: object
          description: Successful Response
      summary: Packs
      tags:
      - catalog
    post:
      operationId: create_pack_draft_v1_catalog_packs_post
      parameters:
      - in: header
        name: Idempotency-Key
        required: true
        schema:
          maxLength: 128
          minLength: 1
          title: Idempotency-Key
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PackCreateRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Create Pack Draft V1 Catalog Packs Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Pack Draft
      tags:
      - catalog
  /v1/catalog/packs/entries:
    post:
      operationId: add_pack_entry_v1_catalog_packs_entries_post
      parameters:
      - in: header
        name: Idempotency-Key
        required: true
        schema:
          maxLength: 128
          minLength: 1
          title: Idempotency-Key
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PackEntryRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Add Pack Entry V1 Catalog Packs Entries Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Add Pack Entry
      tags:
      - catalog
  /v1/catalog/packs/publish-requests:
    post:
      operationId: create_pack_publish_request_v1_catalog_packs_publish_requests_post
      parameters:
      - in: header
        name: Idempotency-Key
        required: true
        schema:
          maxLength: 128
          minLength: 1
          title: Idempotency-Key
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PackPublishRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Create Pack Publish Request V1 Catalog Packs Publish
                  Requests Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Pack Publish Request
      tags:
      - catalog
  /v1/catalog/packs/publish-requests/{request_id}/fulfill:
    post:
      operationId: fulfill_pack_publish_v1_catalog_packs_publish_requests__request_id__fulfill_post
      parameters:
      - in: path
        name: request_id
        required: true
        schema:
          title: Request Id
          type: string
      - in: header
        name: Idempotency-Key
        required: true
        schema:
          maxLength: 128
          minLength: 1
          title: Idempotency-Key
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PackFulfillRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Fulfill Pack Publish V1 Catalog Packs Publish Requests  Request
                  Id  Fulfill Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Fulfill Pack Publish
      tags:
      - catalog
  /v1/catalog/packs/{pack_id}/{version}/export:
    get:
      operationId: export_pack_v1_catalog_packs__pack_id___version__export_get
      parameters:
      - in: path
        name: pack_id
        required: true
        schema:
          title: Pack Id
          type: string
      - in: path
        name: version
        required: true
        schema:
          title: Version
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Export Pack
      tags:
      - catalog
...
  /v1/catalog/refs/refresh:
    post:
      operationId: refresh_ref_v1_catalog_refs_refresh_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRefRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Refresh Ref V1 Catalog Refs Refresh Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Refresh Ref
      tags:
      - catalog
...